Update JsonSerializer.php
ADD: functionality to handle references of objects. It stores every object handle within an array and compares it during the further serialization steps. If the object is already stored, it will be serialized to a reference. The number of the reference is the index of the object within the array. It should be the same order like unserialization.